refactor(cli): split tool startup into cohesive modules
