+++ packages/agent-cli/agent-cli.cabal
@@ -97,6 +97,12 @@ library
                     , Agent.CLI.Runtime.Orchestration.Session
                     , Agent.CLI.Runtime.Orchestration.Startup
                     , Agent.CLI.Runtime.Orchestration.Tools
+                    , Agent.CLI.Runtime.Orchestration.Tools.Request
+                    , Agent.CLI.Runtime.Orchestration.Tools.Model
+                    , Agent.CLI.Runtime.Orchestration.Tools.Collaboration
+                    , Agent.CLI.Runtime.Orchestration.Tools.Scratch
+                    , Agent.CLI.Runtime.Orchestration.Tools.Mcp
+                    , Agent.CLI.Runtime.Orchestration.Tools.HostHooks
                     , Agent.CLI.Runtime.Orchestration.Types
                     , Agent.CLI.Runtime.HistorySource
                     , Agent.CLI.Runtime.Persistence
